requirements.txt
streamlit==1.31.0
torch==2.2.0
torchvision==0.17.0
Pillow==10.2.0
numpy==1.26.4
datasets==2.16.0